requirements.txt
cv2
flask
flask_cors
imaginairy
openai
python-dotenv